Attributes viewer/editor (#11)
New function `attributes_window` allows examining the attributes of the axes and plots within a figure. Editing certain attributes, such as axis labels, titles, various boolean settings is also possible.

This is done using GtkObservables widgets. Many of these do not work for `Observable{Any}`, which is what Makie uses mostly, so some changes to GtkObservables will be needed to make this work for more attributes.